club_charter_expert_github/club_charter_expert/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-docx>=1.1.0
pdfplumber>=0.10.3
pypdf>=4.0.0
